LANG_003_A put edit files from src to the correct directories [CS, DE]
