+++ lang-ui/src/main/java/at/tugraz/sw20asd/lang/ui/controllers/EditVocab.java
@@ -1,22 +1,25 @@
-package at.tugraz.sw20asd.lang.ui.controllers;
 
+
+package at.tugraz.sw20asd.lang.ui.controllers;
 import at.tugraz.sw20asd.lang.EntryDto;
-import at.tugraz.sw20asd.lang.VocabularyDetailDto;
 import at.tugraz.sw20asd.lang.ui.dataaccess.VocabularyAccess;
+import at.tugraz.sw20asd.lang.VocabularyDetailDto;
+
+
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.concurrent.Task;
 import javafx.concurrent.Worker;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextField;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
+import javafx.scene.control.*;
+import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
-import javafx.scene.paint.Color;
+import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 
 import java.io.IOException;
 import java.net.URL;
@@ -25,11 +28,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 
+import javafx.scene.paint.Color;
+
 public class EditVocab extends VBox {
     @FXML
     private VBox phrase_list;
-    @FXML
-    private VBox translation_list;
     @FXML
     private Button add_btn;
     @FXML
@@ -52,15 +55,20 @@ public class EditVocab extends VBox {
 
     private Task<VocabularyDetailDto> getVocabsTask;
     private Task<Integer> edittask1;
-    private Task<Integer> edittask2;
     private VocabularyAccess vocab;
     private VocabularyDetailDto voc;
-    private int origin_scene;
+    private Mode origin_scene;
 
     private long id;
     FXMLLoader loader = new FXMLLoader();
 
-    public EditVocab(VocabularyAccess vocab, long id, int origin_scene) {
+    enum Mode {
+        OVERVIEW,
+        EDIT,
+        TEST,
+    }
+
+    public EditVocab(VocabularyAccess vocab, long id, Mode origin_scene) {
         this.vocab = vocab;
         this.voc = voc;
         this.id = id;
@@ -78,7 +86,6 @@ public class EditVocab extends VBox {
 
 
     public void initialize() {
-        getVocabsGroup();
         user_info.setVisible(false);
 
         ObservableList<String> languages =
@@ -90,6 +97,7 @@ public class EditVocab extends VBox {
         language_map.put("German", Locale.GERMAN);
         language_map.put("English", Locale.ENGLISH);
 
+        getVocabsGroup();
 
         from_choice.valueProperty().addListener((observableValue, oldValue, newValue) -> {
             if (!to_choice.getSelectionModel().isEmpty() &&
@@ -112,19 +120,27 @@ public class EditVocab extends VBox {
         });
 
         add_btn.setOnAction(new EventHandler<ActionEvent>() {
-
             public void handle(ActionEvent event) {
+                HBox entry = new HBox();
+                entry.setMinHeight(30);
+                entry.setAlignment(Pos.TOP_CENTER);
                 TextField phrase = new TextField();
                 phrase.setText("");
+                Label label = new Label("");
+                label.setMinWidth(20);
                 phrase.setId("phrase" + phrase_field_list.size());
+                setMargin(entry, new Insets(5, 0, 0, 0));
                 phrase_field_list.add(phrase);
-                phrase_list.getChildren().add(phrase);
+                phrase_list.getChildren().add(entry);
+                entry.getChildren().add(phrase);
+                entry.getChildren().add(label);
 
                 TextField translation = new TextField();
                 translation.setText("");
                 translation.setId("translation" + translation_field_list.size());
                 translation_field_list.add(translation);
-                translation_list.getChildren().add(translation);
+                setMargin(translation, new Insets(0, 0, 0, 20));
+                entry.getChildren().add(translation);
             }
         });
 
@@ -136,11 +152,11 @@ public class EditVocab extends VBox {
                 } else if (!checkEntries()) {
                     updateUserInformation("entry_missing");
                 } else if (from_choice.getSelectionModel().isEmpty()
-                        || to_choice.getSelectionModel().isEmpty()){
+                        || to_choice.getSelectionModel().isEmpty()) {
                     updateUserInformation("missing_selection");
-                } else if (from_choice.getSelectionModel().getSelectedItem().equals(to_choice.getSelectionModel().getSelectedItem())){
+                } else if (from_choice.getSelectionModel().getSelectedItem().equals(to_choice.getSelectionModel().getSelectedItem())) {
                     updateUserInformation("equal_lang");
-                } else{
+                } else {
                     sendEditCommand();
                     user_info.setVisible(true);
                 }
@@ -150,7 +166,7 @@ public class EditVocab extends VBox {
 
         return_btn.setOnAction(new EventHandler<ActionEvent>() {
             public void handle(ActionEvent event) {
-                if (origin_scene == 1) {
+                if (origin_scene == Mode.OVERVIEW) {
                     OverviewVocabs overview = new OverviewVocabs(vocab);
                     getScene().setRoot(overview);
                 } else {
@@ -168,14 +184,8 @@ public class EditVocab extends VBox {
             case "edited_vocab":
                 user_info.setText("Vocabulary edited!");
                 break;
-            case "amount_missing":
-                user_info.setText("Please fill out the amount");
-                break;
             case "entry_missing":
                 user_info.setText("Please leave no textfields empty");
-                break;
-            case "empty_textfields":
-                user_info.setText("Please check leave no textfields empty");
                 break;
             case "category":
                 user_info.setText("Please fill out the Name of your Vocabulary group");
@@ -190,53 +200,79 @@ public class EditVocab extends VBox {
 
     private void getVocabsGroup() {
         //get Vocabulary group
-        getVocabsTask = new Task<VocabularyDetailDto>() {
-            @Override
-            protected VocabularyDetailDto call() throws Exception {
-                voc = vocab.getVocabulary(id);
-                return voc;
-            }
-        };
+        if (origin_scene == Mode.TEST) {
 
-        getVocabsTask.stateProperty().addListener(((observable, oldValue, newValue) -> {
-            if (newValue == Worker.State.CANCELLED || newValue == Worker.State.FAILED) {
-                Platform.runLater(() -> {
-                    updateUserInformation("");
-                });
-                getVocabsTask.cancel();
-            }
+            category.setText("Pets");
 
-            if (newValue == Worker.State.SUCCEEDED) {
-                if (voc == null) {
-                    Platform.runLater(() -> {
-                        updateUserInformation("No words added");
-                    });
-                } else {
-                    Platform.runLater(() -> {
-                        category.setText(voc.getName());
-                        for (EntryDto e : voc.getEntries()) {
-                            words.add(e);
-                        }
-                        for (int counter = 0; counter < words.size(); counter++) {
-                            TextField phrase = new TextField();
-                            phrase.setText(words.get(counter).getPhrase());
-                            phrase.setId("phrase" + phrase_field_list.size());
-                            phrase_field_list.add(phrase);
-                            phrase_list.getChildren().add(phrase);
-
-                            TextField translation = new TextField();
-                            translation.setText(words.get(counter).getTranslation());
-                            translation.setId("translation" + translation_field_list.size());
-                            translation_field_list.add(translation);
-                            translation_list.getChildren().add(translation);
-                        }
-                    });
+            from_choice.getSelectionModel().select(0);
+            to_choice.getSelectionModel().select(1);
+            words.add(new EntryDto("Hund", "dog"));
+            words.add(new EntryDto("Katze", "cat"));
+            words.add(new EntryDto("Fisch", "fish"));
+            initiateGUI(words);
+        } else {
+            getVocabsTask = new Task<VocabularyDetailDto>() {
+                @Override
+                protected VocabularyDetailDto call() throws Exception {
+                    voc = vocab.getVocabulary(id);
+                    return voc;
                 }
-            }
-        }));
-        Thread th = new Thread(getVocabsTask);
-        th.setDaemon(true);
-        th.start();
+            };
+
+            getVocabsTask.stateProperty().addListener(((observable, oldValue, newValue) -> {
+                if (newValue == Worker.State.CANCELLED || newValue == Worker.State.FAILED) {
+                    Platform.runLater(() -> {
+                        updateUserInformation("");
+                    });
+                    getVocabsTask.cancel();
+                }
+
+                if (newValue == Worker.State.SUCCEEDED) {
+                    if (voc == null) {
+                        Platform.runLater(() -> {
+                            updateUserInformation("No words added");
+                        });
+                    } else {
+                        Platform.runLater(() -> {
+                            category.setText(voc.getName());
+
+                            from_choice.getSelectionModel().select(0);
+                            to_choice.getSelectionModel().select(1);
+
+                            words.addAll(voc.getEntries());
+                            initiateGUI(words);
+                        });
+                    }
+                }
+            }));
+            Thread th = new Thread(getVocabsTask);
+            th.setDaemon(true);
+            th.start();
+        }
+    }
+
+    private void initiateGUI(List<EntryDto> words) {
+        for (int counter = 0; counter < words.size(); counter++) {
+            HBox entry = new HBox();
+            entry.setMinHeight(30);
+            entry.setAlignment(Pos.TOP_CENTER);
+            TextField phrase = new TextField();
+            phrase.setText(words.get(counter).getPhrase());
+            phrase.setId("phrase" + phrase_field_list.size());
+            Label label = new Label("");
+            label.setMinWidth(20);
+            setMargin(entry, new Insets(5, 0, 0, 0));
+            phrase_field_list.add(phrase);
+            phrase_list.getChildren().add(entry);
+            entry.getChildren().add(phrase);
+            entry.getChildren().add(label);
+
+            TextField translation = new TextField();
+            translation.setText(words.get(counter).getTranslation());
+            translation.setId("translation" + translation_field_list.size());
+            translation_field_list.add(translation);
+            entry.getChildren().add(translation);
+        }
     }
 
     private void sendEditCommand() {
@@ -276,6 +312,13 @@ public class EditVocab extends VBox {
                 if (id != -1) {
                     Platform.runLater(() -> {
                         updateUserInformation("edited_vocab");
+                        if (origin_scene == Mode.OVERVIEW) {
+                            OverviewVocabs overview = new OverviewVocabs(vocab);
+                            getScene().setRoot(overview);
+                        } else {
+                            EditVocabSelection editvocabs = new EditVocabSelection(vocab);
+                            getScene().setRoot(editvocabs);
+                        }
                     });
                 } else {
                     Platform.runLater(() -> {
@@ -303,8 +346,7 @@ public class EditVocab extends VBox {
     private boolean checkEntries() {
         for (int counter = 0; counter < phrase_field_list.size(); counter++) {
             if ((!phrase_field_list.get(counter).getText().isEmpty() && translation_field_list.get(counter).getText().isEmpty()) ||
-                 (phrase_field_list.get(counter).getText().isEmpty() && !translation_field_list.get(counter).getText().isEmpty()) ||
-                 (phrase_field_list.get(counter).getText().isEmpty() && translation_field_list.get(counter).getText().isEmpty())){
+                    (phrase_field_list.get(counter).getText().isEmpty() && !translation_field_list.get(counter).getText().isEmpty())) {
                 return false;
             }
         }
