+++ lang-ui/src/main/resources/edit.fxml
@@ -1,11 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.shape.*?>
-<?import javafx.scene.text.*?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.shape.Line?>
+<?import javafx.scene.text.Font?>
 
+<?import javafx.scene.control.ComboBox?>
 <fx:root alignment="TOP_CENTER" prefHeight="600.0" prefWidth="500.0" style="-fx-border-style: solid inside; -fx-border-color: grey; -fx-border-radius: 0; -fx-border-width: 10; -fx-background-color: white;" type="VBox" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="at.tugraz.sw20asd.lang.ui.controllers.EditVocab">
 
     <AnchorPane fx:id="anchor_pane" prefHeight="602.0" prefWidth="800.0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0" AnchorPane.topAnchor="0">
@@ -23,17 +29,6 @@
                 </font>
             </Label>
             <TextField fx:id="category" layoutX="114.0" layoutY="155.0" prefHeight="26.0" prefWidth="299.0" />
-
-<!--            <Label fx:id="from_label_1" layoutX="65.0" layoutY="256.0" prefHeight="17.0" prefWidth="145.0" text="From:">-->
-<!--                <font>-->
-<!--                    <Font name="Abyssinica SIL" size="16.0" />-->
-<!--                </font></Label>-->
-
-<!--            <Label fx:id="to_label_1" layoutX="269.0" layoutY="256.0" prefHeight="17.0" prefWidth="151.0" text="To:">-->
-<!--                <font>-->
-<!--                    <Font name="Abyssinica SIL" size="16.0" />-->
-<!--                </font></Label>-->
-
             <ComboBox fx:id="from_choice" layoutX="62.0" layoutY="192.0" prefHeight="26.0" prefWidth="145.0" />
             <ComboBox fx:id="to_choice" layoutX="268.0" layoutY="192.0" prefHeight="26.0" prefWidth="145.0" />
             
@@ -60,27 +55,27 @@
                         <children>
                            <HBox alignment="TOP_CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="57.0" prefWidth="466.0">
                               <children>
-                                 <VBox fx:id="phrase_list" prefHeight="200.0" prefWidth="100.0">
-                                    <HBox.margin>
-                                       <Insets left="50.0" right="50.0" />
-                                    </HBox.margin>
+                                 <VBox fx:id="phrase_list" alignment="TOP_CENTER" prefHeight="27.0" prefWidth="366.0">
                                     <children>
-                              
-                                                <Label fx:id="from_label" prefHeight="17.0" prefWidth="145.0" text="From:">
-                                              <font>
-                                                  <Font name="Abyssinica SIL" size="16.0" />
-                                              </font>
-                                       </Label>
-                                    </children>
-                                 </VBox>
-                                 <VBox fx:id="translation_list" prefHeight="200.0" prefWidth="100.0">
-                                    <children>
-                              
-                                                <Label fx:id="to_label" prefHeight="17.0" prefWidth="151.0" text="To:">
-                                              <font>
-                                                  <Font name="Abyssinica SIL" size="16.0" />
-                                              </font>
-                                       </Label>
+                                       <HBox alignment="TOP_CENTER" prefHeight="30.0" prefWidth="366.0">
+                                          <children>
+                                    
+                                                      <Label fx:id="from_label" prefHeight="26.0" prefWidth="119.0" text="From:">
+                                                    <font>
+                                                        <Font name="Abyssinica SIL" size="16.0" />
+                                                    </font>
+                                             </Label>
+                                    
+                                                      <Label fx:id="to_label" prefHeight="17.0" prefWidth="151.0" text="To:">
+                                                    <font>
+                                                        <Font name="Abyssinica SIL" size="16.0" />
+                                                    </font>
+                                                <HBox.margin>
+                                                   <Insets left="50.0" />
+                                                </HBox.margin>
+                                             </Label>
+                                          </children>
+                                       </HBox>
                                     </children>
                                     <HBox.margin>
                                        <Insets left="50.0" right="50.0" />
