--- src/main/resources/edit.fxml
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Line?>
-<?import javafx.scene.text.Font?>
-
-<?import javafx.scene.control.ComboBox?>
-<fx:root alignment="TOP_CENTER" prefHeight="600.0" prefWidth="500.0" style="-fx-border-style: solid inside; -fx-border-color: grey; -fx-border-radius: 0; -fx-border-width: 10; -fx-background-color: white;" type="VBox" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="at.tugraz.sw20asd.lang.ui.Controller.EditVocab">
-
-    <AnchorPane fx:id="anchor_pane" prefHeight="602.0" prefWidth="800.0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0" AnchorPane.topAnchor="0">
-        <children>
-
-            <Label fx:id="title" alignment="CENTER" layoutX="63.0" layoutY="75.0" maxWidth="-Infinity" prefHeight="42.0" prefWidth="350" text="Edit vocabulary" textAlignment="CENTER">
-                <font>
-                    <Font name="Abyssinica SIL" size="34.0" />
-                </font></Label>
-
-
-            <Label fx:id="category_label" layoutX="64.0" layoutY="155.0" text="Name:">
-                <font>
-                    <Font name="Abyssinica SIL" size="16.0" />
-                </font>
-            </Label>
-            <TextField fx:id="category" layoutX="114.0" layoutY="155.0" prefHeight="26.0" prefWidth="299.0" />
-            <ComboBox fx:id="from_choice" layoutX="62.0" layoutY="192.0" prefHeight="26.0" prefWidth="145.0" />
-            <ComboBox fx:id="to_choice" layoutX="268.0" layoutY="192.0" prefHeight="26.0" prefWidth="145.0" />
-            
-            <Button fx:id="return_btn" layoutX="63.0" layoutY="467.0" text="Return">
-                <font>
-                    <Font name="Abyssinica SIL" size="16.0" />
-                </font></Button>
-
-            <Button fx:id="submit_btn" layoutX="340.0" layoutY="467.0" text="Submit">
-                <font>
-                    <Font name="Abyssinica SIL" size="16.0" />
-                </font></Button>
-            <Line endX="171.1016387939453" endY="-0.8284531831741333" layoutX="245.0" layoutY="130.0" startX="-180.19125366210938" startY="-0.8284209966659546" strokeMiterLimit="20.0" />
-            <Label fx:id="user_info" layoutX="64.0" layoutY="496.0" prefHeight="42.0" prefWidth="353.0" text="User Information">
-                <font>
-                    <Font size="14.0" />
-                </font>
-            </Label>
-         <ScrollPane hbarPolicy="NEVER" layoutX="1.0" layoutY="231.0" prefHeight="230.0" prefWidth="479.0">
-            <content>
-             <AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="0.0" minWidth="0.0" prefWidth="462.0">
-                  <children>
-                     <VBox alignment="TOP_CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="1000.0" prefWidth="466.0">
-                        <children>
-                           <HBox alignment="TOP_CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="57.0" prefWidth="466.0">
-                              <children>
-                                 <VBox fx:id="phrase_list" alignment="TOP_CENTER" prefHeight="27.0" prefWidth="366.0">
-                                    <children>
-                                       <HBox alignment="TOP_CENTER" prefHeight="30.0" prefWidth="366.0">
-                                          <children>
-                                    
-                                                      <Label fx:id="from_label" prefHeight="26.0" prefWidth="119.0" text="From:">
-                                                    <font>
-                                                        <Font name="Abyssinica SIL" size="16.0" />
-                                                    </font>
-                                             </Label>
-                                    
-                                                      <Label fx:id="to_label" prefHeight="17.0" prefWidth="151.0" text="To:">
-                                                    <font>
-                                                        <Font name="Abyssinica SIL" size="16.0" />
-                                                    </font>
-                                                <HBox.margin>
-                                                   <Insets left="50.0" />
-                                                </HBox.margin>
-                                             </Label>
-                                          </children>
-                                       </HBox>
-                                    </children>
-                                    <HBox.margin>
-                                       <Insets left="50.0" right="50.0" />
-                                    </HBox.margin>
-                                 </VBox>
-                              </children>
-                           </HBox>
-                           <Button fx:id="add_btn" mnemonicParsing="false" text="+" />
-                        </children>
-                     </VBox>
-                  </children>
-               </AnchorPane>
-            </content>
-         </ScrollPane>
-
-        </children>
-    </AnchorPane>
-
-</fx:root>
